[refactor] CoverLetterContent 컴포넌트/훅 분리 및 클린업
